+++ server/models/file_list_item.go
@@ -7,7 +7,8 @@ type FileListItem struct {
 	FileName     string
 	CreateByName string
 	CreateByID   string
-	ChannelName	 string
+	ChannelName  string
 	CreateAt     int64
 	Size         int64
+	HasPreviewImage   bool
 }