+++ webapp/src/action_types.js
@@ -5,3 +5,4 @@ export const CLOSE_ROOT_MODAL = pluginId + '_close_root_modal';
 export const LOAD_FILES = pluginId + "_load_files";
 export const SET_ERROR = pluginId + "_set_error";
 export const SET_CONFIG = pluginId + "_set_config";
+export const LOAD_EXTENSIONS = pluginId + "_load_extensions";